stt_server/requirements.txt
faster-whisper>=1.1.1
kafka-python>=2.2.4
google-cloud-storage>=3.1.0
